split to handler
